requirements.txt
selenium
webdriver-manager
Pillow
reportlab